full_version/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
openpyxl==3.1.2
pandas==2.1.4
sqlalchemy==2.0.23
pyodbc==5.0.1
pillow==10.0.0
requests==2.31.0
crewai==0.70.1
langchain>=0.2.16,<0.3.0
openai==1.35.0